requirements.txt
streamlit
python-dotenv
requests
httpx
openai
